Moves PivotItem to separate folder
